ai: clean up headers; merge "magic" creatures
