+++ meson.build
@@ -129,7 +129,6 @@ sources = [
   'src/game/level.c',
   'src/game/lot.c',
   'src/game/music.c',
-  'src/game/objects/ai/alligator.c',
   'src/game/objects/ai/ape.c',
   'src/game/objects/ai/bacon_lara.c',
   'src/game/objects/ai/baldy.c',
@@ -151,7 +150,6 @@ sources = [
   'src/game/objects/ai/statue.c',
   'src/game/objects/ai/torso.c',
   'src/game/objects/ai/trex.c',
-  'src/game/objects/ai/vole.c',
   'src/game/objects/ai/wolf.c',
   'src/game/objects/boat.c',
   'src/game/objects/bridge.c',
