+++ src/game/objects/ai/bacon_lara.h
@@ -2,8 +2,6 @@
 
 #include "global/types.h"
 
-#include <stdint.h>
-
 void BaconLara_Setup(OBJECT_INFO *obj);
 void BaconLara_Initialise(int16_t item_num);
 void BaconLara_Control(int16_t item_num);
