+++ src/game/objects/ai/bear.c
@@ -8,7 +8,38 @@
 #include "global/types.h"
 #include "global/vars.h"
 
-BITE_INFO g_BearHeadBite = { 0, 96, 335, 14 };
+#define BEAR_CHARGE_DAMAGE 3
+#define BEAR_SLAM_DAMAGE 200
+#define BEAR_ATTACK_DAMAGE 200
+#define BEAR_PAT_DAMAGE 400
+#define BEAR_TOUCH 0x2406C
+#define BEAR_ROAR_CHANCE 80
+#define BEAR_REAR_CHANCE 768
+#define BEAR_DROP_CHANCE 1536
+#define BEAR_REAR_RANGE SQUARE(WALL_L * 2) // = 4194304
+#define BEAR_ATTACK_RANGE SQUARE(WALL_L) // = 1048576
+#define BEAR_PAT_RANGE SQUARE(600) // = 360000
+#define BEAR_RUN_TURN (5 * PHD_DEGREE) // = 910
+#define BEAR_WALK_TURN (2 * PHD_DEGREE) // = 364
+#define BEAR_EAT_RANGE SQUARE(WALL_L * 3 / 4) // = 589824
+#define BEAR_HITPOINTS 20
+#define BEAR_RADIUS (WALL_L / 3) // = 341
+#define BEAR_SMARTNESS 0x4000
+
+typedef enum {
+    BEAR_STROLL = 0,
+    BEAR_STOP = 1,
+    BEAR_WALK = 2,
+    BEAR_RUN = 3,
+    BEAR_REAR = 4,
+    BEAR_ROAR = 5,
+    BEAR_ATTACK1 = 6,
+    BEAR_ATTACK2 = 7,
+    BEAR_EAT = 8,
+    BEAR_DEATH = 9,
+} BEAR_ANIM;
+
+static BITE_INFO m_BearHeadBite = { 0, 96, 335, 14 };
 
 void Bear_Setup(OBJECT_INFO *obj)
 {
@@ -184,7 +215,7 @@ void Bear_Control(int16_t item_num)
 
         case BEAR_ATTACK1:
             if (!item->required_anim_state && (item->touch_bits & BEAR_TOUCH)) {
-                CreatureEffect(item, &g_BearHeadBite, Effect_Blood);
+                CreatureEffect(item, &m_BearHeadBite, Effect_Blood);
                 g_LaraItem->hit_points -= BEAR_ATTACK_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = BEAR_STOP;
