+++ src/game/objects/ai/mutant.c
@@ -12,9 +12,50 @@
 #include "game/sound.h"
 #include "global/vars.h"
 
-BITE_INFO g_WarriorBite = { -27, 98, 0, 10 };
-BITE_INFO g_WarriorRocket = { 51, 213, 0, 14 };
-BITE_INFO g_WarriorShard = { -35, 269, 0, 9 };
+#define FLYER_CHARGE_DAMAGE 100
+#define FLYER_LUNGE_DAMAGE 150
+#define FLYER_PUNCH_DAMAGE 200
+#define FLYER_PART_DAMAGE 100
+#define FLYER_WALK_TURN (PHD_DEGREE * 2) // = 364
+#define FLYER_RUN_TURN (PHD_DEGREE * 6) // = 1092
+#define FLYER_POSE_CHANCE 80
+#define FLYER_UNPOSE_CHANCE 256
+#define FLYER_WALK_RANGE SQUARE(WALL_L * 9 / 2) // = 21233664
+#define FLYER_ATTACK1_RANGE SQUARE(600) // = 360000
+#define FLYER_ATTACK2_RANGE SQUARE(WALL_L * 5 / 2) // = 6553600
+#define FLYER_ATTACK3_RANGE SQUARE(300) // = 90000
+#define FLYER_ATTACK_RANGE SQUARE(WALL_L * 15 / 4) // = 14745600
+#define FLYER_TOUCH 0x678
+#define FLYER_BULLET1 1
+#define FLYER_BULLET2 2
+#define FLYER_FLYMODE 4
+#define FLYER_TWIST 8
+#define FLYER_HITPOINTS 50
+#define FLYER_RADIUS (WALL_L / 3) // = 341
+#define FLYER_SMARTNESS 0x7FFF
+
+#define WARRIOR2_SMARTNESS 0x2000
+
+enum FLYER_ANIM {
+    FLYER_EMPTY = 0,
+    FLYER_STOP = 1,
+    FLYER_WALK = 2,
+    FLYER_RUN = 3,
+    FLYER_ATTACK1 = 4,
+    FLYER_DEATH = 5,
+    FLYER_POSE = 6,
+    FLYER_ATTACK2 = 7,
+    FLYER_ATTACK3 = 8,
+    FLYER_AIM1 = 9,
+    FLYER_AIM2 = 10,
+    FLYER_SHOOT = 11,
+    FLYER_MUMMY = 12,
+    FLYER_FLY = 13,
+};
+
+static BITE_INFO m_WarriorBite = { -27, 98, 0, 10 };
+static BITE_INFO m_WarriorRocket = { 51, 213, 0, 14 };
+static BITE_INFO m_WarriorShard = { -35, 269, 0, 9 };
 
 void Mutant_Setup(OBJECT_INFO *obj)
 {
@@ -230,7 +271,7 @@ void Mutant_FlyerControl(int16_t item_num)
         case FLYER_ATTACK1:
             if (item->required_anim_state == FLYER_EMPTY
                 && (item->touch_bits & FLYER_TOUCH)) {
-                CreatureEffect(item, &g_WarriorBite, Effect_Blood);
+                CreatureEffect(item, &m_WarriorBite, Effect_Blood);
                 g_LaraItem->hit_points -= FLYER_LUNGE_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = FLYER_STOP;
@@ -240,7 +281,7 @@ void Mutant_FlyerControl(int16_t item_num)
         case FLYER_ATTACK2:
             if (item->required_anim_state == FLYER_EMPTY
                 && (item->touch_bits & FLYER_TOUCH)) {
-                CreatureEffect(item, &g_WarriorBite, Effect_Blood);
+                CreatureEffect(item, &m_WarriorBite, Effect_Blood);
                 g_LaraItem->hit_points -= FLYER_CHARGE_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = FLYER_RUN;
@@ -250,7 +291,7 @@ void Mutant_FlyerControl(int16_t item_num)
         case FLYER_ATTACK3:
             if (item->required_anim_state == FLYER_EMPTY
                 && (item->touch_bits & FLYER_TOUCH)) {
-                CreatureEffect(item, &g_WarriorBite, Effect_Blood);
+                CreatureEffect(item, &m_WarriorBite, Effect_Blood);
                 g_LaraItem->hit_points -= FLYER_PUNCH_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = FLYER_STOP;
@@ -279,10 +320,10 @@ void Mutant_FlyerControl(int16_t item_num)
         case FLYER_SHOOT:
             if (flyer->flags & FLYER_BULLET1) {
                 flyer->flags &= ~FLYER_BULLET1;
-                CreatureEffect(item, &g_WarriorShard, Effect_ShardGun);
+                CreatureEffect(item, &m_WarriorShard, Effect_ShardGun);
             } else if (flyer->flags & FLYER_BULLET2) {
                 flyer->flags &= ~FLYER_BULLET2;
-                CreatureEffect(item, &g_WarriorRocket, Effect_RocketGun);
+                CreatureEffect(item, &m_WarriorRocket, Effect_RocketGun);
             }
             break;
