+++ src/game/objects/ai/raptor.c
@@ -7,7 +7,34 @@
 #include "game/random.h"
 #include "global/vars.h"
 
-BITE_INFO g_RaptorBite = { 0, 66, 318, 22 };
+#define RAPTOR_ATTACK_RANGE SQUARE(WALL_L * 3 / 2) // = 2359296
+#define RAPTOR_BITE_DAMAGE 100
+#define RAPTOR_CHARGE_DAMAGE 100
+#define RAPTOR_CLOSE_RANGE SQUARE(680) // = 462400
+#define RAPTOR_DIE_ANIM 9
+#define RAPTOR_LUNGE_DAMAGE 100
+#define RAPTOR_LUNGE_RANGE SQUARE(WALL_L * 3 / 2) // = 2359296
+#define RAPTOR_ROAR_CHANCE 256
+#define RAPTOR_RUN_TURN (4 * PHD_DEGREE) // = 728
+#define RAPTOR_TOUCH 0xFF7C00
+#define RAPTOR_WALK_TURN (1 * PHD_DEGREE) // = 182
+#define RAPTOR_HITPOINTS 20
+#define RAPTOR_RADIUS (WALL_L / 3) // = 341
+#define RAPTOR_SMARTNESS 0x4000
+
+typedef enum {
+    RAPTOR_EMPTY = 0,
+    RAPTOR_STOP = 1,
+    RAPTOR_WALK = 2,
+    RAPTOR_RUN = 3,
+    RAPTOR_ATTACK1 = 4,
+    RAPTOR_DEATH = 5,
+    RAPTOR_WARNING = 6,
+    RAPTOR_ATTACK2 = 7,
+    RAPTOR_ATTACK3 = 8,
+} RAPTOR_ANIM;
+
+static BITE_INFO m_RaptorBite = { 0, 66, 318, 22 };
 
 void Raptor_Setup(OBJECT_INFO *obj)
 {
@@ -119,7 +146,7 @@ void Raptor_Control(int16_t item_num)
             tilt = angle;
             if (item->required_anim_state == RAPTOR_EMPTY && info.ahead
                 && (item->touch_bits & RAPTOR_TOUCH)) {
-                CreatureEffect(item, &g_RaptorBite, Effect_Blood);
+                CreatureEffect(item, &m_RaptorBite, Effect_Blood);
                 g_LaraItem->hit_points -= RAPTOR_LUNGE_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = RAPTOR_STOP;
@@ -130,7 +157,7 @@ void Raptor_Control(int16_t item_num)
             tilt = angle;
             if (item->required_anim_state == RAPTOR_EMPTY && info.ahead
                 && (item->touch_bits & RAPTOR_TOUCH)) {
-                CreatureEffect(item, &g_RaptorBite, Effect_Blood);
+                CreatureEffect(item, &m_RaptorBite, Effect_Blood);
                 g_LaraItem->hit_points -= RAPTOR_CHARGE_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = RAPTOR_RUN;
@@ -141,7 +168,7 @@ void Raptor_Control(int16_t item_num)
             tilt = angle;
             if (item->required_anim_state == RAPTOR_EMPTY
                 && (item->touch_bits & RAPTOR_TOUCH)) {
-                CreatureEffect(item, &g_RaptorBite, Effect_Blood);
+                CreatureEffect(item, &m_RaptorBite, Effect_Blood);
                 g_LaraItem->hit_points -= RAPTOR_BITE_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = RAPTOR_STOP;
